MO-11650: Fixed issue in the reorient mapping when longitude is close to pi/2 or -pi/2
In point_distort() method after the call to gan_cameraf_equirect_xy2latlong(), if the longitude value is too close to pi/2 or -pi/2 then you get errors in the final result. My suspicioin is it relates to the q.z = cos(longitude)*cLat; calculation that is extremely close to 0.0 (in the bug it is -1.39246330e-08 for the pixel column that displays the artefact).